refactor(apps): Phase 3 — migrate uploads and rooms to Zod codecs
Async, cross-converter converters onto the codec pattern via factories
closing over the orchestrator. Behaviour is unchanged — rooms.tests.ts
(both directions, incl. isPartial) and the golden snapshots still pass.

- Add `mappedDecodeAsync`: reproduces `transformMappedData`'s string- and
  (async) function-field branches plus the `_unmappedProperties_` bucket
  and deep-clone isolation, so the function-field converters can drop the
  `transformMappedData` dependency.
- `createUploadsCodec(orch)`: async `decode` resolving room/user/visitor
  through their converters; sync `encode` reverse. Class delegates via
  `z.decodeAsync` / `z.encode`.
- `createRoomCodec(orch)`: async `decode` (convertRoom — creator/visitor/
  contact/department/closedBy/servedBy/responseBy/parentRoom via converters,
  secure fields, RoomType via the shared enum codec) and `encode` (the
  non-partial `convertAppRoom`, reading models directly). `convertRoomRaw`
  becomes a standalone `decodeRoomRaw`. The class keeps the `isPartial`
  variant, which the codec interface can't express.
