+++ apps/meteor/app/apps/server/converters/codecs/mappedData.ts
@@ -63,6 +63,51 @@ export function mappedEncode(fieldMap: FieldMap): (app: Loose) => Loose {
 }
 
 /**
+ * A map whose entries are either a source property name (string) or a function that derives the
+ * target value from the (cloned) source data — the same two entry kinds the legacy
+ * `transformMappedData` supports, minus the nested `{ from, map, list }` form.
+ */
+export type AsyncFieldMap = Record<string, string | ((data: Record<string, any>) => unknown | Promise<unknown>)>;
+
+/**
+ * Rocket.Chat -> Apps-Engine transform for a map mixing string renames and (possibly async)
+ * derived-value functions. This reproduces the string and function branches of the legacy
+ * `transformMappedData` exactly:
+ *
+ * - the input is deep-cloned up front, so functions may freely mutate the clone (e.g. `delete`
+ *   fields they consume) and the app can never mutate the stored document;
+ * - string entries copy the source field to the target name when defined and always drop the source
+ *   key from the bucket;
+ * - function entries receive the clone and set the target only when they return a defined value
+ *   (functions are responsible for deleting any source keys they consume, as before);
+ * - everything left in the clone becomes `_unmappedProperties_`.
+ */
+export async function mappedDecodeAsync(data: Loose, map: AsyncFieldMap): Promise<Loose> {
+	const clone: Loose = structuredClone(data);
+	const result: Loose = {};
+
+	for (const [to, from] of Object.entries(map)) {
+		if (typeof from === 'function') {
+			const value = await from(clone);
+
+			if (typeof value !== 'undefined') {
+				result[to] = value;
+			}
+		} else {
+			if (typeof clone[from] !== 'undefined') {
+				result[to] = clone[from];
+			}
+
+			delete clone[from];
+		}
+	}
+
+	result._unmappedProperties_ = clone;
+
+	return result;
+}
+
+/**
  * Builds a Zod codec from a plain string field map, using {@link mappedDecode} / {@link mappedEncode}.
  *
  * The endpoints are typed with `z.custom` so no runtime validation is added yet (behaviour-preserving);
