+++ apps/meteor/tests/unit/app/apps/server/codecs/mappedData.spec.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 import { describe, it } from 'mocha';
 import * as z from 'zod';
 
-import { createMappedCodec } from '../../../../../../app/apps/server/converters/codecs/mappedData';
+import { createMappedCodec, mappedDecodeAsync } from '../../../../../../app/apps/server/converters/codecs/mappedData';
 
 describe('createMappedCodec', () => {
 	const codec = createMappedCodec({
@@ -51,5 +51,47 @@ describe('createMappedCodec', () => {
 		const original = { _id: 'r1', name: 'admin', scope: 'Users', extra: 1 };
 
 		expect(z.encode(codec, z.decode(codec, original))).to.deep.equal(original);
+	});
+});
+
+describe('mappedDecodeAsync', () => {
+	it('handles string renames, sync and async derived fields, and buckets the rest', async () => {
+		const result = await mappedDecodeAsync(
+			{ _id: 'r1', t: 'c', u: { _id: 'u1' }, leftover: 42 },
+			{
+				id: '_id',
+				type: (data) => {
+					const { t } = data;
+					delete data.t;
+					return t === 'c' ? 'channel' : t;
+				},
+				creator: async (data) => {
+					const { u } = data;
+					delete data.u;
+					return u ? { id: u._id } : undefined;
+				},
+			},
+		);
+
+		expect(result).to.deep.equal({
+			id: 'r1',
+			type: 'channel',
+			creator: { id: 'u1' },
+			_unmappedProperties_: { leftover: 42 },
+		});
+	});
+
+	it('omits derived fields that return undefined and does not mutate the input', async () => {
+		const input = { _id: 'r1', nested: { a: 1 } };
+		const result = await mappedDecodeAsync(input, {
+			id: '_id',
+			creator: () => undefined,
+		});
+
+		expect(result).to.deep.equal({ id: 'r1', _unmappedProperties_: { nested: { a: 1 } } });
+		expect(result).to.not.have.property('creator');
+
+		(result as any)._unmappedProperties_.nested.a = 999;
+		expect(input.nested.a).to.equal(1);
 	});
 });
